+++ python/fast_mlsirm/__init__.py
@@ -30,6 +30,7 @@ from .scaling import lsr_rankings as lsr_rankings, ilsr_rankings as ilsr_ranking
 from .scaling import lsr_top1 as lsr_top1, ilsr_top1 as ilsr_top1
 from .scaling import circular_triads as circular_triads, CircularTriadsResult as CircularTriadsResult
 from .scaling import kendall_u as kendall_u, KendallUResult as KendallUResult
+from .scaling import elo_rating as elo_rating, EloResult as EloResult
 from .cdm import fit_cdm as fit_cdm, CdmFit as CdmFit, fit_gdina as fit_gdina, GdinaFit as GdinaFit, validate_q_matrix as validate_q_matrix, QMatrixValidation as QMatrixValidation, gdina_wald_selection as gdina_wald_selection, WaldModelSelection as WaldModelSelection, fit_ho_cdm as fit_ho_cdm, HoCdmFit as HoCdmFit, fit_ho_gdina as fit_ho_gdina, HoGdinaFit as HoGdinaFit, fit_seq_gdina as fit_seq_gdina, SeqGdinaFit as SeqGdinaFit, fit_seq_gdina_qr as fit_seq_gdina_qr, SeqGdinaQrFit as SeqGdinaQrFit
 from .mixture import fit_mixture as fit_mixture, MixtureFit as MixtureFit
 from .crm import fit_crm as fit_crm, CrmFit as CrmFit
@@ -249,6 +250,8 @@ __all__ = [
     "CircularTriadsResult",
     "kendall_u",
     "KendallUResult",
+    "elo_rating",
+    "EloResult",
     "circle_arc_middle_anchor",
     "CircleArcResult",
     "fit_response_times",